made a utility function for InReview issues and actions + refactored
