+++ gui/types/storyTypes.ts
@@ -72,3 +72,9 @@ export interface SetupModelModalProps {
   openModal: boolean;
   setOpenModel: React.Dispatch<React.SetStateAction<boolean>>;
 }
+
+export interface StoryInReviewIssue {
+  title: string | null;
+  description: string | null;
+  actions: { label: string; link: string }[];
+}
